requirements.txt
setuptools==41.2.0
Flask==1.0.2
flask_cors==3.0.8
PyYAML==5.1.2
pyzmq==18.1.0
